+++ src/cljs/cxengage_javascript_sdk/modules/interaction.cljs
@@ -1,19 +1,18 @@
 (ns cxengage-javascript-sdk.modules.interaction
   (:require-macros [cljs.core.async.macros :refer [go go-loop]]
                    [lumbajack.macros :refer [log]]
-                   [cljs-sdk-utils.macros :refer [def-sdk-fn]])
+                   [cxengage-javascript-sdk.domain.macros :refer [def-sdk-fn]])
   (:require [cljs.core.async :as a]
             [cljs.spec.alpha :as s]
             [cxengage-javascript-sdk.internal-utils :as iu]
             [cxengage-javascript-sdk.state :as state]
-            [cljs-sdk-utils.specs :as specs]
-            [cljs-sdk-utils.errors :as err]
-            [cljs-sdk-utils.protocols :as pr]
+            [cxengage-javascript-sdk.domain.specs :as specs]
+            [cxengage-javascript-sdk.domain.protocols :as pr]
             [cxengage-javascript-sdk.interaction-management :as int]
-            [cljs-sdk-utils.errors :as e]
+            [cxengage-javascript-sdk.domain.errors :as e]
             [cxengage-javascript-sdk.pubsub :as p]
-            [cljs-sdk-utils.interop-helpers :as ih]
-            [cljs-sdk-utils.topics :as topics]
+            [cxengage-javascript-sdk.domain.interop-helpers :as ih]
+            [cxengage-javascript-sdk.domain.topics :as topics]
             [cxengage-javascript-sdk.domain.rest-requests :as rest]))
 
 (s/def ::generic-interaction-fn-params