+++ src/cljs/cxengage_javascript_sdk/modules/twilio.cljs
@@ -1,14 +1,14 @@
 (ns cxengage-javascript-sdk.modules.twilio
   (:require-macros [cljs.core.async.macros :refer [go go-loop]]
-                   [cljs-sdk-utils.macros :refer [def-sdk-fn]]
+                   [cxengage-javascript-sdk.domain.macros :refer [def-sdk-fn]]
                    [lumbajack.macros :refer [log]])
   (:require [cljsjs.paho]
             [cljs.core.async :as a]
             [cxengage-javascript-sdk.state :as state]
-            [cljs-sdk-utils.interop-helpers :as ih]
-            [cljs-sdk-utils.protocols :as pr]
-            [cljs-sdk-utils.errors :as e]
-            [cljs-sdk-utils.topics :as topics]
+            [cxengage-javascript-sdk.domain.interop-helpers :as ih]
+            [cxengage-javascript-sdk.domain.protocols :as pr]
+            [cxengage-javascript-sdk.domain.errors :as e]
+            [cxengage-javascript-sdk.domain.topics :as topics]
             [cxengage-javascript-sdk.domain.rest-requests :as rest]
             [cxengage-javascript-sdk.pubsub :as p]
             [cljs.spec.alpha :as s]))